Delete native methods from java.lang.Thread
'getStackTrace0' is only used in asyncGetStackTrace() which is deleted as well.
Thread.getStackTrace()  is replaced by PlatformThreads.getStackTrace(). So it's not
reachable.

'getNextThreadIdOffset()' is only used in ThreadIdentifiers private class which is
substituted in JDKs better than JDK 19 (which includes 21). So won't be reachable either.

Fixes: #28

This is a partial backport of:
"Introudce a mode to unconditionally include classes" https://github.com/oracle/graal/commit/369f0ff8
to address https://github.com/oracle/graal/issues/9672

Only take the java.lang.Thread native method deletions as that's sufficient to fix
the issue reported in #28.